+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+## Unreleased
+
+- Fix SES generator to emit KiCad-compatible `routing` section